Using R/RMarkdown to create data frames and maps
This demo file is using R Markdown to save data to a dataframe, then populating an HTML page with the relevant information iteratively , as well as a basic interactive map. This is meant as a proof of concept to ensure that these features, and features built upon them will be of use.

Functions include those to create a standard data frame, appending it with additional info, and creating a map.

Testing is done and presented using the Rosetta Stone Entry. All functions can be broken out into separate scripts for easier manipulation later.

Output is a simplistic HTML file which can be modified extensively, including the map, which is interactive and can have legends, etc. added. For now you can click on the blue markers for museum locations, and the orange dot as the origin.